+++ src/features/bosses/types.ts
@@ -1,0 +1,11 @@
+export type BossCategory = 'world boss' | 'boss' | 'aberration'
+
+export interface Boss {
+  name: string
+  imagePath: string
+  category: BossCategory
+  location: 'Labyrinth' | 'Losomn' | 'Nerud' | 'Root Earth' | 'Yaesha'
+  dungeon: string
+  wikiLinks: string[]
+  id: string
+}