+++ src/features/localstorage/useLocalStorage.ts
@@ -1,8 +1,13 @@
 import { GenericItem } from '@/features/items/types/GenericItem'
 import { useLocalStorage as useLS } from 'usehooks-ts'
+import { BossCategory } from '../bosses/types'
 
 // The type of the database in LocalStorage
 interface LocalStorage {
+  bossTracker: {
+    discoveredBossIds: string[]
+    collapsedBossCategories: BossCategory[]
+  }
   tracker: {
     discoveredItemIds: string[]
     collapsedCategories: Array<GenericItem['category']>
@@ -12,6 +17,10 @@ interface LocalStorage {
 }
 
 const initialValue: LocalStorage = {
+  bossTracker: {
+    discoveredBossIds: [],
+    collapsedBossCategories: [],
+  },
   tracker: {
     discoveredItemIds: [],
     collapsedCategories: [],
@@ -20,6 +29,39 @@ const initialValue: LocalStorage = {
 }
 
 export const useLocalStorage = () => {
+  // ----------------------------------------
+  // Boss Tracker
+  // ----------------------------------------
+
+  const [bossTrackerStorage, setBossTrackerStorage] = useLS<
+    LocalStorage['bossTracker']
+  >('boss-tracker', initialValue.bossTracker)
+
+  function setDiscoveredBossIds({
+    ids,
+  }: {
+    ids: LocalStorage['bossTracker']['discoveredBossIds']
+  }) {
+    setBossTrackerStorage({ ...bossTrackerStorage, discoveredBossIds: ids })
+  }
+  const discoveredBossIds = bossTrackerStorage.discoveredBossIds
+
+  function setCollapsedBossCategories({
+    categories,
+  }: {
+    categories: LocalStorage['bossTracker']['collapsedBossCategories']
+  }) {
+    setBossTrackerStorage({
+      ...bossTrackerStorage,
+      collapsedBossCategories: categories,
+    })
+  }
+  const collapsedBossCategories = bossTrackerStorage.collapsedBossCategories
+
+  // ----------------------------------------
+  // Item Tracker
+  // ----------------------------------------
+
   const [itemTrackerStorage, setItemTrackerStorage] = useLS<
     LocalStorage['tracker']
   >('item-tracker', initialValue.tracker)
@@ -45,15 +87,25 @@ export const useLocalStorage = () => {
   }
   const collapsedCategories = itemTrackerStorage.collapsedCategories
 
+  // ----------------------------------------
+  // Sorting Preference
+  // ----------------------------------------
+
   const [sortingPreference, setSortingPreference] = useLS<
     LocalStorage['sortingPreference']
   >('sorting-preference', initialValue.sortingPreference)
 
   return {
+    collapsedBossCategories,
+    setCollapsedBossCategories,
+    discoveredBossIds,
+    setDiscoveredBossIds,
+
     collapsedCategories,
     setCollapsedCategories,
     discoveredItemIds,
     setDiscoveredItemIds,
+
     sortingPreference,
     setSortingPreference,
   }